Stop sanitizing HTML during DOM manipulation
Based on the fact that this script only manipulates existing HTML
(that may or may not include user-generated content) that has been
generated elsewhere, it *does not* fall into the area of responsibility
of this script to guard against XSS attacks.

A benefit of this insight is a drop from 28.6 KB to 7.31 KB in bundle size.