Fix slugging when nesting accordions. (#18)
The `wfaccordionGroup`-constructer selects all children `js-accordion`
elements to build its accordions. This leads to a bug in the slugging
mechanism, when accordions are nested. In the generation of the
outermost `accordionGroup`, the `triggerElement` has already been
cleaned of attributes so when the inner `accordionGroup` is generated,
the slugging defaults to building the slug from the trigger-content,
because there is no `id`-attribute anymore to generate the slug from.
This PR fixes this bug by filtering the selected `js-accordion`-elements
to direct children of the `accordionGroup`.